requirements.txt
python-telegram-bot[job-queue]==20.7
python-dotenv==1.1.1
playwright==1.53.0
Pillow==10.2.0
nest-asyncio==1.5.7